+++ src/update.rs
@@ -64,12 +64,6 @@ impl Release {
         self.assets
             .iter()
             .find(|asset| {
-                // Skip .sha(1|256|512) checksum files
-                let extension = asset.name.split('.').last().unwrap_or_default();
-                if extension.starts_with("sha") {
-                    return false;
-                }
-
                 asset.name.contains(target)
                     || (asset.name.contains(OS) && asset.name.contains(ARCH))
                         && if let Some(i) = identifier {
@@ -81,6 +75,9 @@ impl Release {
             .cloned()
     }
 }
+
+/// Function type for determining the asset among multiple matches
+pub type AssetMatchFn = fn(&Release, &str, Option<&str>) -> Option<ReleaseAsset>;
 
 /// Updates to a specified or latest release
 pub trait ReleaseUpdate {
@@ -103,6 +100,9 @@ pub trait ReleaseUpdate {
     fn identifier(&self) -> Option<String> {
         None
     }
+
+    /// Optional function for determining the asset among multiple matches
+    fn asset_match_fn(&self) -> Option<AssetMatchFn>;
 
     /// Name of the binary being updated
     fn bin_name(&self) -> String;
@@ -212,11 +212,13 @@ pub trait ReleaseUpdate {
             }
         };
 
-        let target_asset = release
-            .asset_for(&target, self.identifier().as_deref())
-            .ok_or_else(|| {
-                format_err!(Error::Release, "No asset found for target: `{}`", target)
-            })?;
+        let target_asset = if let Some(asset_match_fn) = self.asset_match_fn() {
+            asset_match_fn(&release, &target, self.identifier().as_deref())
+        } else {
+            release.asset_for(&target, self.identifier().as_deref())
+        }.ok_or_else(|| {
+            format_err!(Error::Release, "No asset found for target: `{}`", target)
+        })?;
 
         let prompt_confirmation = !self.no_confirm();
         if self.show_output() || prompt_confirmation {