+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## [unreleased]
 ### Added
+- add `UpdateBuilder::asset_match_fn` to enable custom logic for release asset matching
 ### Changed
-- skip `sha` (`sha1`, `sha256`, `sha512`) checksum files when determining the release asset 
 ### Removed
 
 ## [0.40.0]